# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Campaign Nexus",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rule-systems/{system_id}/schema/{sheet_type}": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Schema",
        "operationId": "get_schema_api_v1_rule_systems__system_id__schema__sheet_type__get",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          },
          {
            "name": "sheet_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sheet Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Schema Api V1 Rule Systems  System Id  Schema  Sheet Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-systems/{system_id}/layout/{sheet_type}": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Layout",
        "operationId": "get_layout_api_v1_rule_systems__system_id__layout__sheet_type__get",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          },
          {
            "name": "sheet_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sheet Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Layout Api V1 Rule Systems  System Id  Layout  Sheet Type  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-systems/{system_id}/validate": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Validate Doc",
        "operationId": "validate_doc_api_v1_rule_systems__system_id__validate_post",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-systems/{system_id}/facets": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Facets",
        "operationId": "get_facets_api_v1_rule_systems__system_id__facets_get",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacetDefOut"
                  },
                  "title": "Response Get Facets Api V1 Rule Systems  System Id  Facets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-systems/{system_id}/conditions": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Conditions",
        "operationId": "get_conditions_api_v1_rule_systems__system_id__conditions_get",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConditionOut"
                  },
                  "title": "Response Get Conditions Api V1 Rule Systems  System Id  Conditions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rule-systems/{system_id}/travel": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Travel Table",
        "description": "Paces, conveyances and terrain multipliers \u2014 the travel planner's vocabulary.",
        "operationId": "get_travel_table_api_v1_rule_systems__system_id__travel_get",
        "parameters": [
          {
            "name": "system_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Travel Table Api V1 Rule Systems  System Id  Travel Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/stat-blocks": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Blocks",
        "operationId": "list_blocks_api_v1_campaigns__campaign_id__stat_blocks_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "sheet_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sheet Type"
            }
          },
          {
            "name": "rule_system_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rule System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StatBlockOut"
                  },
                  "title": "Response List Blocks Api V1 Campaigns  Campaign Id  Stat Blocks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Block",
        "operationId": "create_block_api_v1_campaigns__campaign_id__stat_blocks_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatBlockCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatBlockOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/stat-blocks/{block_id}": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Block",
        "operationId": "get_block_api_v1_campaigns__campaign_id__stat_blocks__block_id__get",
        "parameters": [
          {
            "name": "block_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Block Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatBlockOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "rules"
        ],
        "summary": "Update Block",
        "operationId": "update_block_api_v1_campaigns__campaign_id__stat_blocks__block_id__put",
        "parameters": [
          {
            "name": "block_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Block Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatBlockUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatBlockOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/monsters": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "List Monsters",
        "operationId": "list_monsters_api_v1_campaigns__campaign_id__monsters_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "facet1_num_gte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facet1 Num Gte"
            }
          },
          {
            "name": "facet1_num_lte",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facet1 Num Lte"
            }
          },
          {
            "name": "facet1_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facet1 Text"
            }
          },
          {
            "name": "facet2_text",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facet2 Text"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MonsterOut"
                  },
                  "title": "Response List Monsters Api V1 Campaigns  Campaign Id  Monsters Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/monsters/import": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Import Bestiary",
        "operationId": "import_bestiary_api_v1_campaigns__campaign_id__monsters_import_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          },
          {
            "name": "system_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "System Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/monsters/export": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Export Bestiary",
        "operationId": "export_bestiary_api_v1_campaigns__campaign_id__monsters_export_get",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Export Bestiary Api V1 Campaigns  Campaign Id  Monsters Export Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/monsters/import-json": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Import Bestiary Json",
        "operationId": "import_bestiary_json_api_v1_campaigns__campaign_id__monsters_import_json_post",
        "parameters": [
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Import Bestiary Json Api V1 Campaigns  Campaign Id  Monsters Import Json Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/monsters/{monster_id}": {
      "get": {
        "tags": [
          "rules"
        ],
        "summary": "Get Monster",
        "operationId": "get_monster_api_v1_campaigns__campaign_id__monsters__monster_id__get",
        "parameters": [
          {
            "name": "monster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monster Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/campaigns/{campaign_id}/monsters/{monster_id}/variant": {
      "post": {
        "tags": [
          "rules"
        ],
        "summary": "Create Variant",
        "operationId": "create_variant_api_v1_campaigns__campaign_id__monsters__monster_id__variant_post",
        "parameters": [
          {
            "name": "monster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monster Id"
            }
          },
          {
            "name": "campaign_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Campaign Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonsterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConditionOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "description"
        ],
        "title": "ConditionOut"
      },
...
      "FacetDefOut": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "type"
        ],
        "title": "FacetDefOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidateRequest": {
        "properties": {
          "sheet_type": {
            "type": "string",
            "title": "Sheet Type"
          },
          "doc": {
            "additionalProperties": true,
            "type": "object",
            "title": "Doc"
          }
        },
        "type": "object",
        "required": [
          "sheet_type",
          "doc"
        ],
        "title": "ValidateRequest"
      },
      "ValidateResult": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors"
          },
          "derived": {
            "additionalProperties": true,
            "type": "object",
            "title": "Derived"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "errors",
          "derived"
        ],
        "title": "ValidateResult"
      },
...
    }
  }
}